feat: Backwards compat helicone.features config
this commit updates the `helicone` config with a new `helicone.features`
field to reduce confusion around the previous configuration option of
two separate boolean flags for `authentication` and `observability`.

the `features` field is an enum which prevents invalid combinations
(such as enabling observability without enabling auth).

this change is backwards compataible, the old flags will still
deserialize and will be mapped to the new enum as appropriate